Added basic installer
This commit marks the beginning of a plan to completely overhaul the way
that FitNexus is distributed and installed. The plan is, roughly:

- Create an installer script that builds a new FitNexus project from
  scratch (bin/fitnexus). This installer will be the primary frontend
  to FitNexus.
- Create a gemspec so FitNexus can be built and distributed as a gem.
  This gem will depend on the gems that FitNexus needs (namely Rsel
  and Cukable).

Once the istaller is ready and is capable of generating a new FitNexus
project from scratch, nearly everything else in this repository can go
away, including the stub FitNesseRoot, the mastiffe and rsel submodules,
symlinks, and so on.